+++ packages/eslint-plugin/rules/use-recommended-components.js
@@ -20,6 +20,7 @@ const ALLOWLIST = {
 			'Badge',
 			'Calendar',
 			'Card',
+			'CheckboxControl',
 			'Collapsible',
 			'CollapsibleCard',
 			'ControlWithError',